Configure the build system for MPI

+++ source/main.cpp
@@ -3,14 +3,20 @@
 #include <string>
 #include "revision.hpp"
 
+#ifdef MPI_ENABLE
+#include <mpi.h>
+#endif
+
 int main(int argc, char* argv[])
 {
   std::string lockfilename="hydro3.lock";
+#ifndef MPI_ENABLE
   if (file_exists(lockfilename)) {
     std::cout << "Error: " + lockfilename +
         " exists. Another instance seems to be running." << std::endl;
     return -1;
   }
+#endif
 
   std::ofstream lockfile(lockfilename);
   lockfile.close();
@@ -69,6 +75,29 @@ int main(int argc, char* argv[])
     exitcode=-1;
   }
 
-  remove_file(lockfilename);
+  if (file_exists(lockfilename)) {
+    remove_file(lockfilename);
+  }
+  
+#ifdef MPI_ENABLE
+  MPI_Init(NULL, NULL);
+
+  int world_size;
+  MPI_Comm_size(MPI_COMM_WORLD, &world_size);
+
+  int world_rank;
+  MPI_Comm_rank(MPI_COMM_WORLD, &world_rank);
+  
+  char processor_name[MPI_MAX_PROCESSOR_NAME];
+  int name_len;
+  MPI_Get_processor_name(processor_name, &name_len);
+  
+  printf("Hello world from processor %s, rank %d"
+         " out of %d processors\n",
+         processor_name, world_rank, world_size);
+
+  MPI_Finalize();
+#endif
+  
   return exitcode;
 }
